Replace package.yaml with edna.cabal
Problem: we want to be able to build edna package with cabal and
that requires edna.cabal file. We can store it in the repo and
somehow ensure it's up-to-date or can generate it (e. g. using
a script) whenever we do `cabal build`.

Solution: both options have some overhead and `hpack` doesn't seem
to be very beneficial here because we have just one package.
So we simply don't use `hpack` and store only edna.cabal in the repo.