Add endpoints to query ranges of projects and methodologies
Problem: library page needs to show existing projects and
methodologies, but currently it has no way to query them.

Solution: add 2 endpoints to do that.
Both of them take pagination and sorting parameters, but they are stubs
for now, we'll take care of proper types and implementation later.
They return all data that should be shown according to the current
design.
For projects we have some extra data, for methodologies we currently
don't have any extra data, but may have it in future.