Change the root test runner to hspec
Problem: we are using `tasty` as the root test runner, but we implement
all tests using `hspec`. There is a redundancy because these libraries
are quite equivalent and both can be used to group tests and run them.
Thus we have unnecessary dependencies and a level of indirection –
`tasty-hspec` which converts `hspec` Specs to `tasty` trees
imprecisely.

Solution: we need to choose which lib to use. Our consensus is that
both are ok. I've chosen `hspec` because it's already a dependency
of `servant-swagger`, which is kinda silly, but we can't change it.
So all `tasty` packages are not used anymore. `hspec-dicover` is now
used to find tests.