Initialize state machine testing
Problem: we want to test Edna in various scenarios. Currently we just
need to test `parseFile` against various states of Edna, later we will
have more actions to test.

Solution: here we use a known technique called state machine testing.
Hedgehog has builtin support for this kind of tests.
We start with a simple model with just one command that doesn't even
update state (`ParseFile`).
Since it's hard to generate Excel files in tests, we don't test
`parseFile` as is, we test a version of it with already parsed Excel
file.
We check that `parseFie'` doesn't modify state and that it returns
correct result.
Result computation is implemented the same way as in tested code,
except that we fetch IDs from `EdnaState` rather than DB.
Later we will have actions that upload data to DB and these tests
will become more useful.