Move and implement Target endpoints
Problem: currently Target endpoints are not implemented and
Target functionality doesn't match the structure of modules we
want to follow.

Solution:
1. Create Edna.Library.* modules following our common structure.
The whole Library functionality will be in that hierarchy.
2. Implement both Target endpoints.

API has changed a bit: from now on timestamps will be returned as
LocalTime (JSON strings).

SqlId has been moved into Edna.Util because it's used not only at
the level of Web. And now it stores Word32 instead of Word to match
types in DB schema.