Add logging to server endpoints
Problem: currently there is no logging in the server so it's hard to
understand what's going on when you want to debug the server or just
check what it does.

Solution: we start with adding logging to server endpoints using
logging middleware provided by `wai-extra`.
We modify it to log to `stderr` because that's a more standard
destination for logs, at least in Serokell.
We also introduce a very simple logging config with just 3 modes.
That should be sufficient for our current needs.