Store new data about sub-experiments
Problem: there is new functionality that was recently added to Edna
design. Now sub-experiments have names and one of them is the primary
one. Currently we don't store this data in DB.

Solution: add a new field to the `sub_experiment` table and its
Haskell ADT.
Also add a new table that bounds an experiment with its primary
sub-experiment. We could also put this data into `experiment` table,
but then we would have circular references which are not convenient to
deal with. So I think this solution is better.