Add isEnabled flag to MeasurementResp
Problem: we return points on plot as a list of MeasurementResp where
each one is a pair of numbers. But we should also distinguish enabled
and disabled points somehow. Both groups of points should be shown
somehow, but they should be shown differently.

Solution: add a boolean flag to MeasurementResp.