Implement Dashboard getters
Problem: DashboardEndpoints are not implemented.

Solution: we start with implementing getter endpoints.
Primitive DB queries are defined in DB.Query, Service layer
does some additional logic and conversion.
There are some changes in the API:
1. Methodology ID inside `ExperimentResp` is now optional because
for some experiments there may be no methodology (if it's deleted).
2. `SubExperimentResp` doesn't contain `isDefault` flag anymore.
ID of the primary sub-experiment is returned as part of
`ExperimentResp` instead.