Use ordered maps in FileContents
Problem: we are using unordered HashMaps in FileContents.
In tests we sometimes care about the order because it affects assigned
IDs. For HashMaps the order is unspecified and it complicates
development of tests.

Solution: the only advantage of HashMaps is performance, but here
volumes of data shouldn't be big enough for the difference to be
noticeable. So we use ordered Maps instead.
Apart from that we define `sortFileSummary` function and sort
FileSummary that we compute. This way it's easier to check for
FileSummary correctness in tests.

Tests are updated to match the new ordering.