Throw an error on attempt to delete unknown methodology
Problem: currently `deleteMethodology` silently does nothing
when someone tries to delete unknown methodologies.
Because of that we may miss a bug. And it may be confusing to receive
successful response when nothing is deleted.

Solution: throw an error and return an error when someone tries
to delete an unknown test methodology.