Disable default nginx page in frontend docker image
Problem: nginx docker image has a default virtual server configured in
file `default.conf`, which is served when you use `localhost` hostname
as an address, and it overrides the virtual server we use for edna in
docker-compose deployment.

Solution: use `default.conf.template` instead of `edna.conf.template`
for the config template name, so that it replaces the `default.conf`
config.