Use named volume instead of a local directory for db in docker-compose
Problem: when we use a local directory as a volume, docker creates a
root-owned directory inside the project. It becomes harder to clean it
up, because it requires root rights, and git complains about directory
contents being unavailable.

Solution: use a named volume instead, which can be managed with `docker
volume` commands.