Allow to configure server with command line arguments and env variables
Problem:
Currently edna-server is configured using a yaml file, so one needs to pass config.yaml with all options. It's more or less ok for MVP, but it's not sufficiently flexible.

Solution:
Add functionality to pass command line arguments and use environment variables. The most priority has cla, then env and then config.
Also there are changes in docker-compose.yaml, so now we can use .env file to specify db port, user, password and db name. But if we do we also have to export this variables to override default one in backend configuration.