Add endpoint with metrics for prometheus
Problem: We want to know what's going on inside edna-server, at least whether it feels well (in broad sense). We typically use Prometheus for this kind of task.

Solution: Add endpoint with metrics. To collect it we use metrics provided by ghc. Also for now we don't need 'health' endpoint anymore, so it's deleted.
To know is Edna alive we can use 'metrics' endpoint.