Update haskell to work with python script
Problem: currently we use hardcoded values to compute 4pl.
Instead of this we need to run python script which will
compute it.

Solution: refactor analyse4PL function to run python script.
It convert data to json and get json output from stdout.
Additionally insertExperiments function was changed to work
with list of experiments.