Add pagination for Library getters
Problem: the number of projects, compounds and other entities stored
in Edna may be quite big sometimes. In this case we don't want to
transfer all entities to frontend at once, it may take quite a lot of
time to load a page.
Also it should be possible to sort entities in tables. We can do it on
frontend if we return all items in one call, but if we return chunks of
data we need to know sorting order.

Solution: add pagination to library getters that return lists.
We use `servant-pagination` to achieve that. We get sorting as a bonus.
An alternative is to use `servant-util` that provides sorting and
pagination capabilities. I chose `servant-pagination` because it's on
Hackage, it's mentioned in official Servant docs and this way we can
compare it with `servant-util` that we used before.

Some notes:
1. We still fetch all items from DB and use `applyRange` to filter
them. It's better than what we had because we don't transfer more
data than requested, but it's not the most efficient solution.
I want to start with this solution and we can optimize it as a separate
task.
2. There is a plenty of boilerplate in definitions of `get*s` methods.
I made a number of attempts to avoid it, but didn't succeed. One of the
reasons is that `ExtractRange` and `PutRange` range classes are not
exported so we can't write certain generic functions. Besides, this
generic code becomes extremely complex, so I decided to resort to
boilerplate for now. Perhaps some of this boilerplate should be
encapsulated inside `servant-pagination` itself. If we decide that
it's a good library, we can contribute it there.