Avoid division by zero in `computeMeanIC50`
Problem: `computeMeanIC50` divides sum of IC50 values by their number
without checking that their number is non-zero. And nothing guarantees
that it can't be zero. So we may produce infinity or Nan or runtime
error which is not good.

Solution: if there are no IC50 values, set their mean to 0.