Add simple unit tests for analysis
Problem: recently added tests (for upload with auto outliers)
indirectly cover new functionality of `analyse4PL`,
but they are quite high-level and may fail for many reasons.
So it's not obvious which tests cover `analyse4PL` functionality
(especially proper data exchange with Python) and if something
goes wrong in this functionality it may be hard to figure out the
source of the bug.

Solution: add a module with unit tests specifically for `analyse4PL`
and check simple things there.