Use typed-process instead of process
Problem: we are using the `process` library to execute `python`
process, but there is `typed-process` based on `process` which is
clearly superior. Specifically:
1. Much better and more detailed documentation (nice README).
2. The description of `process` refers to `typed-process` saying that
it's a better replacement.
3. Uses `ByteString` instead of `String` in some places, which is saner
and more efficient.
4. Has a slightly different API which looks nicer, as can be seen in
this patch.

Solution: rewrite most of `callPythonAnalysis` to use `typed-process`
functions. Drop explicit `process` dependency completely (we still
transitively depend on it of course).