Print a message when Python is misconfigured
Problem: some functionality in edna-server implicitly relies on proper
configuration of Python environment. If something is misconfigured,
we won't notice it until this functionality is used (e. g. a new
file is uploaded and we perform 4PL analysis). Moreover, it may
be not obvious that the error is caused by misconfiguration of Python.

Solution: check environment immediately on start. To do that, we call
4PL analysis script with empty data. If everything is setup correctly,
it should simply print empty JSON list and exit successfully. If it
fails for any reason, most likely something is misconfigured and we
print a detailed message in this case.
We don't crash because many features still work and I think it would
be unfair to crash, I often want to run `edna-server` and don't care
about ability to call analysis (and don't want to bother with `poetry`
just for that).

Note that when we have more files we'll probably want to test them as
well. We'll think about better design when we have more main files.