Add `run` command to backend's Makefile
Problem: currently it's a bit non-trivial to run edna-server.
It's not enough to just do `stack exec` or `cabal run`, you need
to make sure all Python stuff is available and we are using `poetry`
for that.

Solution: to make it possible to run `edna-server` without manually
entering `poetry shell`, we define `run` command similar to the `test`
command.