Make one test more robust
Problem: in one `DashboardSpec` test we hard-code the expected IC50
value. It's value computed by external code (in Python) and doesn't
have to be precisely equal to the hard-coded one. Minor changes in
Python code may change it slightly, but it will still be valid.
Moreover, it's not clear whether this value comes from.

Solution: compare with the value we receive from `getExperiments` call
with the value computed by `analyse4PLOne` computed for expected
measurements.