Add docker restart policy
Problem:
In docker doc we have following notice: "Docker recommends that you use restart policies, and avoid using process managers to start containers."
So in case Edna is running somewhere on a server with Docker and accidentally this server fall down, after server restart Edna app will not be launched.

Solution:
Set docker restart policy to `always`, so with this setting (if docker daemon starts automatically) Edna will be launched automatically.
Here is description of this setting from Docker docs:
"always - Always restart the container if it stops. If it is manually stopped, it is restarted only when Docker daemon restarts or the container itself is manually restarted."