Set LineBuffering for stderr on start
Problem: stderr has NoBuffering by default. We are using `hPutStrLn`
from the `text` package to print log messages and this function
behaves differently depending on the buffering mode of the handler
it writes to. If it's NoBuffering, it prints each character separately
and concurrent writes may interleave printing of the message
arbitrarily.
As result, when many concurrent requests are made, we may see some
garbage in logs.

Solution: set stderr buffering to LineBuffering on start.
It allows us to use `hPutStrLn` and the whole message is printed
atomically.
Another advantage of LineBuferring is that it's likely more efficient
because output is flushed less often. And we print messages with `\n`
anyway, so they should be flushed as soon as they are printed.