Add primary IC50 to ExperimentResp
Problem: in order to show Dashboard table we query ExperimentsResp
first and then for each experiment in the response we query its primary
sub-experiment one-by-one. It's a bit inefficient and we can easily
decrease number of HTTP endpoint calls.

Solution: add a new field to ExperimentResp that holds IC50 computed
for the primary sub-experiment. Experiment getter was already querying
all sub-experiments, so adding this data to response is pretty
staightforward. We extend an existing test to check `erPrimaryIC50` sanity.