Return no mean IC50 when there are no values
Problem: `computeMeanIC50` returns 0 if there are no experiments
with successfully computed IC50. But we already support the ability
to return no mean IC50 and for no successful experiments it seems
to make more sense than 0, because 0 may mean that all IC50s are 0
for example.

Solution: return `Nothing` if there are no successful IC50s instead
of 0.