Do not fetch each sub-experiment to show Dashboard table
Problem: in order to show Dashboard table we fetch sub-experiments
one-by-one, so there are a lot of requests as the number of experiments
grows. We fetch them just to get primary IC50.

Solution: now that primary IC50 is provided by the backend as part
of ExperimentResp, we use this value instead of fetching SubExperiment.
Experiment and ExperimentDto types were extended with a new field.
`ResultDto` now supports the ability to store just one number.