Don't request experiment's subexperiments until a user opens experiment table collapsible
Problem: we request all subexperiments for every experiment in the
experiment table on Dashboard page, though, nobody will ever open them.
This happens because we don't check if a collapsible is visible before
rendering 'ExperimentsCollapse' component.

Solution: add check if experiment's collapse is visible
before rendering a collapsible. It will cause a spinner
for first time we open it.