Add pagination+sorting for Library getters
Problem: the number of projects, compounds and other entities stored
in Edna may be quite big sometimes. In this case we don't want to
transfer all entities to frontend at once, it may take quite a lot of
time to load a page.
Also it should be possible to sort entities in tables. We can do it on
frontend if we return all items in one call, but if we return chunks of
data we need to know sorting order.

Solution: add pagination to library getters that return lists.
We use `servant-util` and `servant-util-beam-pg` to achieve that.