Add pagination for `getExperiments`
Problem: total number of experiments in the system can be enormous
and currently we send all of them to frontend if `getExperiments`
endpoint is used without any filters. It can cause substantial
slow down to page loading.

Solution: basically the same as in the previous commit. As a temporary
limitation we support only `uploadDate` field, other fields will
be supported in a separate task.
Bonus: `StubSortBy` is not needed anymore and was removed.