Attempt to fix caching issues by disabling cache
Problem: incorrect caching headers for index.html might result in
users seeing an old version. What's even worse, an old version of
index.html may refer to an old JS file that is no longer accessible,
in which case blank page will be shown until hard reload.

Solution: to fix that, we include a Cache-Control header. We first
tried setting it to no-cache that still allows the browser to validate
the cache with the upstream server, at the cost of higher traffic to us.
However, it was not enough because browser uses ETag for cache
validation which is built from mtime and the size of the file and since
we are using nix we can't rely on mtime (they seem to be constant).
So if we modify JS code, but don't change the size of index.html,
we will get the same ETag and assume that nothing has changed (which
is wrong).
So we use `no-store` option which means
"The response may not be stored in any cache".

Co-authored-by: Ivan Gromakovskii <gie1994@gmail.com>
Co-authored-by: Zhenya Vinogradov <zhenyavinogradov@gmail.com>