Fix transactions execution
Problem:
We have three functions: `withConnection`, `transact`  and `runPg`. Function `transact` supposed to wrap some code inside transactions, and it’s working almost as expected. If we look at postgresql log we can see that transactions starts with process id `X` and ends with same one, but actions inside transaction has another id `Y`.
So, the problem is that when we call transact we use withConnection and particular actions uses withConnection too, which produces different processes.

Solution:
Store functions to work with connections in `EdnaContext` instead of `ConnPoll`, and then when we need to execute some actions in transaction - override that functions to use single connection.