Support sorting of experiments by methodology name
Problem: it's possible to sort experiments by upload date, but there
are other columns in the table with experiments and it often makes
sense to sort by them as well.

Solution:
1. Add "methodology" to `SortingParamTypesOf ExperimentResp`, thus
making it possible to pass `+methodology` sorting order for example.
2. Update `getExperiments` query to get the whole `TestMethodology`
record and include it into `sortingApp`.
3. Add default sorting by ID to make order of entries with equal
methodology name defined precisly.
4. Methodology name is now included into `ExperimentResp` along with
methodology ID.