Improve error message in our of the test
Problem: if something goes wrong in Library test, `Map.!` operator may
fail and the error message in this case is not very helpful:
```
       uncaught exception: ErrorCall
       Map.!: given key is not an element in the map
       CallStack (from HasCallStack):
         error, called at libraries/containers/containers/src/Data/Map/Internal.hs:627:17 in containers-0.6.2.1:Data.Map.Internal
```
It refers to some library file and doesn't say what key is not found.

Solution: if key is not found, fail using `expectationFailure` and
attach the key to the error.