Support sorting of experiments by compound and target
Problem: we've added sorting of experiments by test methodology,
but there are other columns for which sorting can be useful and
is not supported yet.

Solution: add sorting by compound and target similarly to methodology.
Sorting by IC50 can be supported too, but with some extra tricks.
So it's postponed for future work, let's see whether it will be
requested.