Switch to openapi3
Problem:
We use swagger2 library that follows Swagger specification version 2. For quite a while a newer specification is available and called OpenAPI 3. AFAIK it's sufficiently mature and all reasonable tools can handle it already. We should use it because it's more modern.
Besides, in the openapi3 library ToParamSchema class returns Schema while in swagger2 it returns ParamSchema. ParamSchema has less fields, in particular it seems to have no Description field, so we can't provide a description for response headers for example.

Solution:
Update resolver to 17.10 because of dependencies pain (for example `servant-swagger-ui` was refactored recently to use more general Value type, which allows us to use openapi3 with swagger ui, and updated package included to newer snapshot).
Change `lenses-micro` to `lenses`, because required instances defined for `Control.Lens.At` are transitively visible from `openapi3` and we transitively depend on lens anyway.
The rest of changes is just updating declarations according to new standard and defining some instances.