Fix dashboard table renderring
Problem: after sorting addition to the dashboard table the arrow
we show for sorting is shown below column name and it doesn't look
nice, also doesn't match design.

Solution: make some css hacks to make it look properly. The problem
was caused by the fact that sometimes IC50 values take too much width,
e. g. if the value is grater than 1e6 or something like that.
We force the value to never exceed a certain width by trimming it and
inserting ellipsis when it exceeds the limit.