Add "names" getters to edna-server
Problem: sometimes we need to know names of all entities of certain,
e. g. all projects on upload page, so that the user can select a
project. Currently it can be done by calling `/projects`, but it's a
bit expensive because it also gets more data than necessary.

Solution: add new endpoints to get names of all entities.
For projects we have 2 such getters: all projects in the system
(including projects without any experiments) and only projects with
experiments.
All projects in the system are needed for the Upload page and
all projects with experiments are needed for the Dashboard page.