Add getter of number of experiments
Problem: in order to implement pagination on frontend it would be
convenient to have a quick way to know the total number of experiments
to be shown in dashboard.

Solution: add `deGetExperimentsNumber` endpoint similar to
`deGetExperiments` which returns just the number of experiments
instead of experiments themselves (and lacks sorting and pagination
arguments for obvious reasons).