Provide more license information in some frontend files
Problem: there are some frontend files that we copy-pasted from
somewhere. Their licenses require us to include their copyright notice
and license text, but we don't include them yet. Specifically, these
files are: `Spinner.scss` style and all PTRootUI fonts.

Solution:
1. Refer to the MIT license in `Spinner.scss`.
2. License information about fonts is included into `PTRootUI.scss`
file where the fonts are imported. It seems to be a common approach
permitted by the OFL license and.

Note that we use `/*!` comments, so they are not removed in production
mode.