Add existence checks in logic
Problem:
We had `Word32` as type for IDs and `serial` in our schemas. Postgres defines `serial` as integer with auto increment and range 1 to 2147483647 (but in fact it's just an alias to integer, hence it able to have negative range).
It means, that upper range of `Word32` is twice bigger that it's allowed by postgres.
To avoid some future errors we have tmo options: change `Word32` to `Int32` - but we don't want to have negative ids; the other one - is just add some checks of existence - because sqlErrors appears only in UPDATE commands (in SELECT everything is fine).

Solution:
Add existence checks in logic to avoid using illegal ids in update commands.