Add Dashboard integration tests
Problem:
We've added some tests for Library API where we call HTTP endpoints. But it's not cover all API and how this API compatible with common standards.

Solution:
Add some tests for Dashboard which calls our API as client with some mock data and checks response data and status codes for errors.