Workaround incorrect dashboard table renderring on some platforms
Problem: sometimes sub-experiments in the dashboard table are shown
incorrectly: there is only one column with them, but the idea is to
show two of them in one row.

Solution: we couldn't find a proper solution and the exact reason
for this behavior. Moreover, it happens only on some platforms,
but doesn't happen on MacBooks and iPhone for example.
As a workaround we simply decrease the width of the sub-experiment
plate so that 2 sub-experiments fit into the width with higher
probability.