Remove quotes from `.env` file
Problem: docker-compose seems to pass values in quotes in env-file
in such a way that the underlying application sees this quotes,
as if they were escaped. Apparently it doesn't affect our case because
we explicitly forward all variables using `environment`, but some
changes to the configuration may cause quotes to be passed to
`edna-server` and cause incorrect behavior (e. g. if there are quotes
around MDe links).

Solution: remove all quotes from `.env` file, it doesn't seem to spoil
or break anything, but seems safer this way. Add a comment asking
to be careful with quotes.