View selectors bug fix
Problem: our filters work on all experiments, we load them when we load
the dashboard page.
When you click on the View button, you load only experiments with
specific project/compound/target (it's what `filteredExperimentsQuery`
does). So if you go to dashboard via View, selectors won't know about
most of experiments and they will usually suggest starting a new filter
not knowing that some project has experiments with various compounds
and targets.

Solution: as a quick workaround we define `experimentsQuery` the same
way as `filteredExperimentsQuery`, except that it always loads all
experiments. Then we use it in dashboard to get all experiments.