Change copyright to copyleft in the footer
Problem: our code is licensed on terms of AGPL-3.0 which is a copyleft
license (https://en.wikipedia.org/wiki/Copyleft) and it's common
to use the copyleft instead of the copyright sign in such case.
This way it's clearer that people can distribute and modify
this intellectual property.
However, we ahve the copyright sign in the footer.

Solution: use CSS to rotate the copyright sign. While there is the
copyleft symbol in Unicode, we are not sure how widely it's supported,
so using CSS seems safer (and admitely more tiresome).