Make collapsible depends on element id
Problem: when sorting order in the dashboard table changes,
we show expanded experiments in the same rows as they were before,
but these are different experiments. A more natural approach is to
expand the same experiments (with the same IDs) as were expanded
before, not the same row indices.

Solution: keep track of IDs of experiments and update collapsible
state based on these IDs.